+++ lib/govuk_publishing_components/presenters/contextual_navigation.rb
@@ -16,7 +16,7 @@ module GovukPublishingComponents
         content_item["document_type"] == "simple_smart_answer"
       end
 
-      def html_document_with_parent?
+      def html_publication_with_parent?
         (content_item["document_type"] == "html_publication") && parent_api_path
       end